refactor: ChatCommandService/ChatQueryService를 ChatService로 통합
